Minor fixes for Fortran+NetCDF support.
1) Fortran was not consistently capitalized internally, and was
   generally conflicting with the cmake convention "Fortran" not "FORTRAN".

2) When removing duplicates in the library list, it is important to
   recognize that there are "debug" libs and "optimized" libs, and the
   duplicates should be removed separately from each rather than mixed
   together.  This is masked by higher levels of ecbuild which strip
   out these options.

There remain other harder issues that must eventually be fixed.  E.g.,
The Fortran "debug" library list is empty which is not a supported
capability for cmake.

And of course the higher levels of cmake are removing duplicates
incorrectly as well.  They should use the same reverse-remove-reverse
pattern, and must be cognizant of debug/optimized issues.  I actually
think that ecbuild should _not_ remove duplicates, but maybe there
were good reasons for this.  But the existing practice definitely gets
it wrong for NetCDF prior to the other workarounds in the NetCDF
layer.